backend/requirements.txt
Django==5.2.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==24.1
psycopg2-binary==2.9.10
redis==5.2.1
django-redis==5.4.0
Pillow==11.2.1
gunicorn==22.0.0
django-summernote==0.8.20
bleach==6.1.0 